refactor: extract page stub utilities
